Implement Carl Londahl's factorisation for close p and q
More details on the algorithm can be found here:
https://grocid.net/2017/09/16/finding-close-prime-factorizations/

This PR fixes issue #59.

We first build a lookup table with `b` entries and then calculate `b` entries
to see if they are in the lookup table. This finds factors if our guess for phi
is at most b^2 away from the actual phi, which gives better results than the
Fermat algorithm.  Increasing b increases memory usage and computation time and
finds primes that are less close to each other. Setting a timeout on this
function is a bad idea, since it will build the lookup table and then abort
while it is using the table.